New CLI argument `offsets_history`
This will determine how much history we keep in the `partition_offsets` Register.

Also, got rid of `brokers_timezone`: the timestamps that we will read from `__consumer_offsets` it's "amount of ms since Unix Epoch", that is UTC by definition. So, everything in here can just be `DateTime<Utc>` everywhere.